+++ assets/php/settings.php
@@ -28,9 +28,6 @@ Translation::setLocalesDir(__DIR__ . '/../locales');
 $_moduleName = Translation::of('settings');
 $_moduleLink = 'index.php?site=settings';
 
-if(getenv("H_IP")) $host_ip = getenv("H_IP") . (getenv("H_PORT") != '80' ? ':' . getenv("H_PORT") : '');
-else $host_ip = $_SERVER['SERVER_ADDR'] . ($_SERVER['SERVER_PORT'] != '80' ? ':' . $_SERVER['SERVER_PORT'] : '');
-
 // Public Access Link
 if(isset($_GET['generateToken']) && $_GET['generateToken'] == 'yes' && (getGroupID($loginUserID) == 1 || hasSettingsPublicRight($loginUserID))){
   $now 	 = time();
@@ -235,7 +232,7 @@ else if(isset($_GET['view']) && $_GET['view'] == 'system' && hasSettingsSystemRi
               </tr>
               <tr>
                 <td>'.Translation::of('server_ip').':</td>
-                <td>'.$host_ip.'</td>
+                <td>'.$_SERVERIP.'</td>
               </tr>
               '.$server_output.'
               <tr>
@@ -347,7 +344,7 @@ else if(isset($_GET['view']) && $_GET['view'] == 'system' && hasSettingsSystemRi
 }
 else if(isset($_GET['view']) && $_GET['view'] == 'publicaccess' && hasSettingsPublicRight($loginUserID)){
 
-  $tokenLink = 'http://'.$_SERVER['SERVER_ADDR'].':'.$_SERVER['SERVER_PORT'].'/_public.php?key='.$set['token'];
+  $tokenLink = 'http://'.$_SERVERIP.'/_public.php?key='.$set['token'];
 
   echo '
   <div class="container-xl">